refactor process files functions
