+++ update_local_v1.py
@@ -4,6 +4,7 @@ import json
 from pathlib import Path
 from params import *  # Expects definitions for METADATA, RAW_FILE_TYPES, etc.
 from generate_html import generate_html_from_json
+from auxiliary import process_files_for_summarised
 import pandas as pd
 import numpy as np
 
@@ -21,7 +22,7 @@ def load_sample_tb(file_path):
     Returns:
         dict: Mapping from "Patient ID" to the metadata entry.
     """
-    with open(file_path) as f:
+    with open(file_path, "r") as f:
         data = json.load(f)
     return data
 
@@ -51,12 +52,12 @@ def filter_files(directory):
     for ext in config["summarised_file_types"]:
         bucket_by_ext[ext.lower()] = "summarised"
 
-    file_dict = {
+    file_path_dict = {
         bucket.replace("_file_types", ""): []
         for bucket in config
     }
 
-    # file_dict = {"raw":[], "processed":[], "summarised":[]}
+    # file_path_dict = {"raw":[], "processed":[], "summarised":[]}
 
     for root, _, files in os.walk(directory):
         for file in files:
@@ -66,18 +67,19 @@ def filter_files(directory):
             if bucket is None:
                 continue
             full_path = Path(root) / file
-            file_dict[bucket].append(full_path)
+            file_path_dict[bucket].append(full_path)
 
-    for bucket, files in file_dict.items():
+    for bucket, files in file_path_dict.items():
         if not files:
             print(f" | No files found for {bucket} file types")
          
-    return file_dict
+    return file_path_dict
 
-def process_files(directory, file_dict, file_type, organization, cor_dict):
-    
+def process_files(directory, file_path_dict, file_type, organization, cor_dict):
+    print(f"Processing the {file_type} files")
     file_list = []
-    for full_path in file_dict[file_type]:
+    total_size = 0
+    for full_path in file_path_dict[file_type]:
 
         relative_path = full_path.relative_to(directory)
         file_path = f"./{relative_path.as_posix()}"
@@ -95,199 +97,15 @@ def process_files(directory, file_dict, file_type, organization, cor_dict):
             "directory": file_path,
             "organization": organization
         }
-        # print("Processing the Raw files")
+
         print(f" | {file_path}  ~{file_size}{FILE_SIZE_UNIT}")
 
         # check here to prevent duplicates
         if metadata_dict not in file_list:
-            file_list.append(file_dict)
+            file_list.append(metadata_dict)
 
     print(f" | Total size for these files: {total_size}{FILE_SIZE_UNIT}")
                         
-    return file_list
-
-def process_files_for_raw(directory, file_types, organization, cor_dict):
-    """   
-    Recursively scans the given directory for raw data files whose names end with one of the specified file_types.
-    Each found file is registered in the RO‑Crate with enriched properties.
-    The file’s base name (without extension) is used to look up its metadata record.
-    
-    Args:
-        directory (str): The directory to search.
-        file_types (list): List of file extensions to match.
-        organization (str): Organization that the data files are from, can be modified in params.py
-        cor_dict: The dictionary containing the keys as sample_id and values as patient_id
-    
-    Returns:
-        list: A list of dictionaries summarizing the file details.
-    """
-    file_list = []
-    total_size = 0
-    for root, _, files in os.walk(directory):
-        for file in files:
-            if os.path.splitext(file)[1] in RAW_FILE_TYPES:
-                full_path = os.path.join(root, file)
-                relative_path = os.path.relpath(full_path, directory)
-                file_path = f"./{relative_path}"
-                file_size = round(os.path.getsize(full_path) / CONVERT_FROM_BYTES)
-                total_size += file_size
-
-                sample_name, _ = os.path.splitext(file)  # name of FASTQ/FASTA is sample name
-
-                # establish file name
-                file_dict = {
-                    "file_name": file,
-                    "file_size": file_size, 
-                    "patient_id": cor_dict.get(sample_name, ""),
-                    "sample_id": sample_name,
-                    "directory": file_path,
-                    "organization": organization
-                }
-                # print("Processing the Raw files")
-                print(f" | {file_path}  ~{file_size}{FILE_SIZE_UNIT}")
-
-                # check here to prevent duplicates
-                if file_dict not in file_list:
-                    file_list.append(file_dict)
-
-    if not file_list:
-        print(" | No files found for file types:", file_types)
-    else:
-        print(f" | Total size for these files: {total_size}{FILE_SIZE_UNIT}")
-                        
-    return file_list
-
-
-
-def process_files_for_summarised(directory, file_types, organization, cor_dict):
-    """   
-    Recursively scans the given directory for summary data files whose names end with one of the specified file_types.
-    Each found file is registered in the RO‑Crate with enriched properties.
-    The file’s base name (without extension) is used to look up its metadata record.
-    
-    Args:
-        directory (str): The directory to search.
-        file_types (list): List of file extensions to match.
-        metadata_dict (dict): Dictionary mapping "Patient ID" to metadata entries.
-        crate (ROCrate): The RO‑Crate instance in which to register files.
-        organization (str): Organization that the data files are from, can be modified in params.py
-        cor_dict: The dictionary containing the keys as sample_id and values as patient_id
-    
-    Returns:
-        list: A list of dictionaries summarizing the file details.
-    """
-    file_list = []
-    total_size = 0
-    for root, _, files in os.walk(directory):
-        for file in files:
-            if os.path.splitext(file)[1] in SUMMARISED_FILE_TYPES:
-                if  os.path.splitext(file)[1] == '.csv':                
-                    full_path = os.path.join(root, file)
-                    relative_path = os.path.relpath(full_path, directory)
-                    file_path = f"./{relative_path}"
-                    file_size = round(os.path.getsize(full_path) / CONVERT_FROM_BYTES)
-                    total_size += file_size
-                    
-                    df = pd.read_csv(full_path, index_col=0)
-
-                    file_dict = {
-                        "file_name": file, 
-                        "file_size": file_size, 
-                        "patient_id": list(set([cor_dict.get(sample, "") for sample in list(df.index)])),
-                        "sample_id": list(df.index),
-                        "directory": file_path,
-                        "organization": organization
-                    }
-                    # print("Processing the summarised files")
-                    print(f" | {file_path}  ~{file_size}{FILE_SIZE_UNIT}")
-
-                    # check here to prevent duplpicates
-                    if file_dict not in file_list:
-                        file_list.append(file_dict)
-
-                elif os.path.splitext(file)[1] == '.tsv':
-                    full_path = os.path.join(root, file)
-                    relative_path = os.path.relpath(full_path, directory)
-                    file_path = f"./{relative_path}"
-                    file_size = round(os.path.getsize(full_path) / CONVERT_FROM_BYTES)
-                    total_size += file_size
-                    
-                    df = pd.read_csv(full_path, index_col=0, delimiter="\t")
-
-                    file_dict = {
-                        "file_name": file, 
-                        "file_size": file_size, 
-                        "patient_id": list(set([cor_dict.get(sample, "") for sample in list(df.index)])),
-                        "sample_id": list(df.index),
-                        "directory": file_path,
-                        "organization": organization
-                    }
-                    print(f" | {file_path}  ~{file_size}{FILE_SIZE_UNIT}")
-
-                    # check here to prevent duplpicates
-                    if file_dict not in file_list:
-                        file_list.append(file_dict)
-                        
-    if not file_list:
-        print(" | No files found for file types:", file_types)
-    else:
-        print(f" | Total size for these files: {total_size}{FILE_SIZE_UNIT}")
-    
-    return file_list
-
-
-
-def process_files_for_processed(directory, file_types, organization, cor_dict):
-    """   
-    Recursively scans the given directory for processed data files whose names end with one of the specified file_types.
-    Each found file is registered in the RO‑Crate with enriched properties.
-    The file’s base name (without extension) is used to look up its metadata record.
-    
-    Args:
-        directory (str): The directory to search.
-        file_types (list): List of file extensions to match.
-        metadata_dict (dict): Dictionary mapping "Patient ID" to metadata entries.
-        crate (ROCrate): The RO‑Crate instance in which to register files.
-        organization (str): Organization that the data files are from, can be modified in params.py
-        cor_dict: The dictionary containing the keys as sample_id and values as patient_id
-    
-    Returns:
-        list: A list of dictionaries summarizing the file details.
-    """
-    file_list = []
-    total_size = 0
-    for root, _, files in os.walk(directory):
-        for file in files:
-            if os.path.splitext(file)[1] in PROCESSED_FILE_TYPES:
-                full_path = os.path.join(root, file)
-                relative_path = os.path.relpath(full_path, directory)
-                file_path = f"./{relative_path}"
-                file_size = round(os.path.getsize(full_path) / CONVERT_FROM_BYTES)
-                total_size += file_size
-
-                sample_name, _ = os.path.splitext(file)  # name of FASTQ/FASTA is sample name
-
-                # establish file name
-                file_dict = {
-                    "file_name": file,
-                    "file_size": file_size, 
-                    "patient_id": cor_dict.get(sample_name, ""),
-                    "sample_id": sample_name,
-                    "directory": file_path,
-                    "organization": organization
-                }
-                # print("Processing the Processed files")
-                print(f" | {file_path}  ~{file_size}{FILE_SIZE_UNIT}")
-
-                # check here to prevent duplpicates
-                if file_dict not in file_list:
-                    file_list.append(file_dict)
-
-    if not file_list:
-        print(" | No files found for file types:", file_types)
-    else:
-        print(f" | Total size for these files: {total_size}{FILE_SIZE_UNIT}")
-
     return file_list
 
 
@@ -308,13 +126,15 @@ def generate_json(directory, output_file):
     # Load metadata from the provided metadata file.
     cor_dict = load_sample_tb(SAMPLE_TO_PATIENT)
     organization = ORGANIZATION
-    
+
+    file_path_dict = filter_files(directory)
+
     print(f"\nProcessing raw files ({', '.join(RAW_FILE_TYPES)})")
-    raw_files = process_files_for_raw(directory, RAW_FILE_TYPES, organization, cor_dict)
-    
+    raw_files = process_files(target_directory, file_path_dict, "raw", organization, cor_dict) 
+
     print(f"\nProcessing processed files ({', '.join(PROCESSED_FILE_TYPES)})")
-    processed_files = process_files_for_processed(directory, PROCESSED_FILE_TYPES, organization, cor_dict)
-    
+    processed_files = process_files(target_directory, file_path_dict, "processed", organization, cor_dict) 
+   
     print(f"\nProcessing summarised files ({', '.join(SUMMARISED_FILE_TYPES)})")
     summarised_files = process_files_for_summarised(directory, SUMMARISED_FILE_TYPES, organization, cor_dict)
